Move module interface code into module.rs of the hyperonc
